Add node modules caching to Github workflow
This change adds node modules caching to the Github workflow (pull-request.yml) to speed up CI runs. By using the 'actions/cache' Github action and keying off the 'package-lock.json' file, it's ensured that unnecessary re-installations of dependencies are avoided. If the cache is missed, a debug command to list the state of node modules is run.